requirements.txt
google-genai==1.27.0
openai==1.97.1
python-dotenv==1.1.1
rich==14.1.0